+++ notification-application/src/main/kotlin/gloddy/notification/dto/NotificationGetDto.kt
@@ -1,0 +1,4 @@
+package gloddy.notification.dto
+
+class NotificationGetDto {
+}
