+++ notification-application/src/main/kotlin/gloddy/notification/port/in/NotificationGetUseCase.kt
@@ -1,0 +1,5 @@
+package gloddy.notification.port.`in`
+
+interface NotificationGetUseCase {
+
+}
